SayMyDigit/srcVision/requirements.txt
flask
numpy
mnist
Pillow
tensorflow